Added single timer creation capability
Single timer creation is working, series timer creation is not implemented, yet.
Timer update works in general, but after an update of the timer a manual reload of the web-page is required.
Fix : Live-streams were not unique, this could lead into a broken stream, if multiple instances to one source are connected.